+++ x-pack/plugins/infra/server/lib/alerting/metric_threshold/lib/evaluate_alert.ts
@@ -6,6 +6,7 @@
  */
 
 import { mapValues, first, last, isNaN } from 'lodash';
+import moment from 'moment';
 import { ElasticsearchClient } from 'kibana/server';
 import {
   isTooManyBucketsPreviewException,
@@ -111,6 +112,26 @@ const getMetric: (
 ) {
   const { aggType } = params;
   const hasGroupBy = groupBy && groupBy.length;
+
+  const interval = `${timeSize}${timeUnit}`;
+  const intervalAsSeconds = getIntervalInSeconds(interval);
+  const intervalAsMS = intervalAsSeconds * 1000;
+
+  const to = moment(timeframe ? timeframe.end : Date.now())
+    .add(1, timeUnit)
+    .startOf(timeUnit)
+    .valueOf();
+
+  // We need enough data for 5 buckets worth of data. We also need
+  // to convert the intervalAsSeconds to milliseconds.
+  // TODO: We only need to get 5 buckets for the rate query, so this logic should move there.
+  const minimumFrom = to - intervalAsMS * MINIMUM_BUCKETS;
+
+  const from = roundTimestamp(
+    timeframe && timeframe.start <= minimumFrom ? timeframe.start : minimumFrom,
+    timeUnit
+  );
+
   const searchBody = getElasticsearchMetricQuery(
     params,
     timefield,
@@ -183,7 +204,7 @@ const getValuesFromAggregations = (
 
     if (aggType === Aggregators.COUNT) {
       return buckets.map((bucket) => ({
-        key: bucket.to_as_string,
+        key: bucket.from_as_string,
         value: bucket.doc_count,
       }));
     }
@@ -192,11 +213,28 @@ const getValuesFromAggregations = (
         const values = bucket.aggregatedValue?.values || [];
         const firstValue = first(values);
         if (!firstValue) return null;
-        return { key: bucket.to_as_string, value: firstValue.value };
+        return { key: bucket.from_as_string, value: firstValue.value };
       });
     }
+
+    if (aggType === Aggregators.AVERAGE) {
+      return buckets.map((bucket) => ({
+        key: bucket.key_as_string ?? bucket.from_as_string,
+        value: bucket.aggregatedValue?.value ?? null,
+      }));
+    }
+
+    if (aggType === Aggregators.RATE) {
+      return buckets
+        .map((bucket) => ({
+          key: bucket.key_as_string ?? bucket.from_as_string,
+          value: bucket.aggregatedValue?.value ?? null,
+        }))
+        .filter(dropPartialBuckets(dropPartialBucketsOptions));
+    }
+
     return buckets.map((bucket) => ({
-      key: bucket.key_as_string ?? bucket.to_as_string,
+      key: bucket.key_as_string ?? bucket.from_as_string,
       value: bucket.aggregatedValue?.value ?? null,
     }));
   } catch (e) {
